fix: resolve frontend module export issues and Node.js compatibility
 **Frontend Configuration & Compatibility Fixes:**

**Node.js Compatibility:**
- Updated Node.js from v20.10.0 to v22.19.0 (LTS) for Vite compatibility
- Fixed Vite v7.1.6 requirements and compilation issues

**PostCSS & Tailwind CSS v4 Configuration:**
- Updated postcss.config.js to use '@tailwindcss/postcss' plugin
- Added @tailwindcss/postcss dependency for v4 compatibility
- Updated src/index.css to use new Tailwind v4 import syntax: @import 'tailwindcss'

**TypeScript Module Structure & Exports:**
- Created src/types/index.ts to centralize all type definitions
- Fixed module export conflicts causing 'User' export not found errors
- Separated type definitions from API service logic
- Updated imports across components to use centralized types

**Component Import Updates:**
- AuthContext.tsx: Import User type from types instead of API service
- PromptEditor.tsx: Import Prompt, Variable, CreatePromptData from types
- PromptList.tsx: Import Prompt type from types
- api.ts: Re-export types for backward compatibility

**Dependencies & Build:**
- Added @tailwindcss/postcss package for PostCSS plugin support
- Updated package.json and package-lock.json with new dependencies
- Fixed TypeScript compilation errors and circular dependencies

**Testing & Documentation:**
- Added api-test-guide.md for manual API testing
- Verified both backend (port 3001) and frontend (port 5173) servers running
- Confirmed module loading and type exports working correctly

**Result:**
 Frontend loads without module export errors
 Tailwind CSS v4 styling applies correctly
 TypeScript compilation successful
 Full-stack application ready for user testing
 Epic 1, Story 1 implementation complete and functional
