+++ frontend/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@tailwindcss/postcss": "^4.1.13",
     "@types/react-router-dom": "^5.3.3",
     "axios": "^1.12.2",
     "react": "^19.1.1",
